ynx_bot/requirements.txt
aiogram>=3.0.0
hydra-core>=1.3.2
omegaconf>=2.3.0
poetry